move calcInterval to substrate
